практическое хзадание 5 плюс

+++ MenagerMertricaComputer/AgentMetricaComputer/ClassForInterfeis/HardDriverAgentMetrica.cs
@@ -10,7 +10,7 @@ namespace AgentMetricaComputer
         public int Id { get; set; }
         public int Value { get; set; }
 
-        public long Time { get; set; }
+        public int Time { get; set; }
 
 
     }
